azf/requirements.txt
azure-functions
appdirs==1.4.4
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
distlib==0.3.2
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
pipenv==2021.5.29
psycopg2==2.9.1
PyGreSQL==5.2.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
azure.storage.blob